Don't sort flattened search match sets
Set uses Hash as storage. Hashes enumerate their values in the order
that the corresponding keys were inserted. The latter is in the
language specification. The former may be an implementation detail
(woof).

Consider reimplementing the DanceDatatable.matching_figures_for_*
methods to return Arrays of SearchMatch when order is important.

See contradb/contra#553